requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
dateparser==1.0.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.4
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-binance==0.7.11
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
service-identity==18.1.0
six==1.15.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc2
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
threadpoolctl==2.1.0
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.4.0
